Move some of the live data plotting into it's own class to simply the processing
