+++ cfar.py
@@ -1,6 +1,8 @@
 import numpy as np
 
-from config import get_cfar_params, CFARParams, CfarType
+from config import get_cfar_params
+from configuration.CFARType import CfarType
+from configuration.CFARParams import CFARParams
     
 def cfar_required_cells(cfar_params: CFARParams = get_cfar_params()):
     if cfar_params.cfar_type == CfarType.CASO:
@@ -68,5 +70,28 @@ def leading_edge_cfar_single(data, index_CUT, cfar_params: CFARParams):
     detection = 1 if signal_level > threshold else 0
     
     return detection, threshold
+
+def get_range_bin_for_index(index, bin_size):
+    return index * bin_size
+
+def get_range_bin_for_indexs(indexArray, bin_size):
+    resultArray = None
+    if indexArray is not None and indexArray.size > 0:
+        resultArray = indexArray * bin_size
+    return resultArray
+
+def get_range_bins(bin_size):
+    range_bins = np.arange(1, 513) * bin_size
+    return range_bins
+
+def get_range_bins_for_distance(min_bin_distance, max_bin_distance, bin_size):
+    range_bins = get_range_bins(bin_size)
+    selected_bins = [rb for rb in range_bins if min_bin_distance <= rb <= max_bin_distance]
+    return selected_bins
+
+def get_range_bin_indexes(min_bin_distance, max_bin_distance, bin_size):
+    range_indexs = np.arange(1, 513)
+    selected_indexes = [rb for rb in range_indexs if min_bin_distance <= rb*bin_size <= max_bin_distance]
+    return np.array(selected_indexes)
