+++ constants.py
@@ -1,28 +1,2 @@
-# First experiment, moving forward then back up to 4.25 meters around 10 seconds each way
-# SAVED_CSV_FILE_NAME = "./data_samples/walking_up_back_in_room.csv"
-
-# SAVED_CSV_FILE_NAME = "street_car.csv"
-
-# 4 cars at the begining of the run, driving by quickly. Then 1 at the end.
-# SAVED_CSV_FILE_NAME = "street_car_4cars_break_1car_at_end.csv"
-
-# 2 cars, somewhat evenly spaced moving slowly
-# SAVED_CSV_FILE_NAME = "street_car_2_cars_evenly_spaced.csv"
-
-# 1 car passing by quickly
-# SAVED_CSV_FILE_NAME = "street_car_1_quickly_passing.csv"
-
-# 3 cars swopping by, radar poiting up street at a diagonal
-# SAVED_CSV_FILE_NAME = "street_car2.csv"
-# 3 cars swooping by again
-# SAVED_CSV_FILE_NAME = "street_car3.csv"
-
-#3 cars swooping by nicely - good results, no people walking by in the background
-# SAVED_CSV_FILE_NAME = "street_cars_3driving_byhouse_diagonal.csv"
-
-# SAVED_CSV_FILE_NAME = "./data_samples/street_cars_3driving_byhouse_diagonal.csv"
-
-SAVED_CSV_FILE_NAME = "./data_samples/new_data.csv"
-
 SPEED_LIGHT = 299792458  # Speed of light in m/s
 DIST_BETWEEN_ANTENNAS = 0.00625 # Distance between recievers of radar
