--- object_tracking/detection_patterns.py
@@ -1,48 +1,0 @@
-import numpy as np
-
-# successive detection patterns (1 is a detection, 0 is not a detection)
-object_tracking_straight = np.array((
-    [0,0,0,1,0,0,0,0], 
-    [0,0,0,1,0,0,0,0], 
-    [0,0,0,1,0,0,0,0], 
-    [0,0,0,1,0,0,0,0], 
-    [0,0,0,1,0,0,0,0], 
-    [0,0,0,1,0,0,0,0], 
-    [0,0,0,1,0,0,0,0], 
-    [0,0,0,1,0,0,0,0], 
-    [0,0,0,1,0,0,0,0], 
-    [0,0,0,1,0,0,0,0], 
-    [0,0,0,1,0,0,0,0], 
-    [0,0,0,1,0,0,0,0], 
-    [0,0,0,1,0,0,0,0]))
-
-object_tracking_across = np.array((
-    [1,0,0,0,0,0,0,0], 
-    [0,1,0,0,0,0,0,0], 
-    [0,0,1,0,0,0,0,0], 
-    [0,0,0,1,0,0,0,0], 
-    [0,0,0,0,1,0,0,0], 
-    [0,0,0,0,0,1,0,0], 
-    [0,0,0,0,0,0,1,0], 
-    [0,0,0,0,0,1,0,0], 
-    [0,0,0,0,1,0,0,0], 
-    [0,0,0,1,0,0,0,0], 
-    [0,0,1,0,0,0,0,0], 
-    [0,1,0,0,0,0,0,0], 
-    [1,0,0,0,0,0,0,0]))
-
-object_tracking_across_double = np.array((
-    [1,1,0,0,0,0,0,0], 
-    [0,1,1,0,0,0,0,0], 
-    [0,0,1,1,0,0,0,0], 
-    [0,0,0,1,1,0,0,0], 
-    [0,0,0,0,1,1,0,0], 
-    [0,0,0,0,0,1,1,0],
-    [0,0,0,0,0,0,1,1], 
-    [0,0,0,0,0,0,1,1], 
-    [0,0,0,0,0,1,1,0], 
-    [0,0,0,0,1,1,0,0], 
-    [0,0,0,1,1,0,0,0], 
-    [0,0,1,1,0,0,0,0], 
-    [0,1,1,0,0,0,0,0], 
-    [1,1,0,0,0,0,0,0]))
